# frontend/src/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Ground Truth Curator",
		"version": "1.0.6"
	},
	"paths": {
...
		"/v1/datasets/{datasetName}/curation-instructions": {
			"get": {
				"tags": ["datasets"],
				"summary": "Get Curation Instructions",
				"operationId": "get_curation_instructions_v1_datasets__datasetName__curation_instructions_get",
				"parameters": [
					{
						"name": "datasetName",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Datasetname"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatasetCurationInstructions"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["datasets"],
				"summary": "Put Curation Instructions",
				"operationId": "put_curation_instructions_v1_datasets__datasetName__curation_instructions_put",
				"parameters": [
					{
						"name": "datasetName",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Datasetname"
						}
					},
					{
						"name": "If-Match",
						"in": "header",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"title": "If-Match"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CurationInstructionsUpdate"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatasetCurationInstructions"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/v1/datasets/{datasetName}": {
			"delete": {
				"tags": ["datasets"],
				"summary": "Delete Dataset",
				"description": "Hard-delete a dataset.",
				"operationId": "delete_dataset_v1_datasets__datasetName__delete",
				"parameters": [
					{
						"name": "datasetName",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Datasetname"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": {
										"type": "string"
									},
									"title": "Response Delete Dataset V1 Datasets  Datasetname  Delete"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CurationInstructionsUpdate": {
				"properties": {
					"instructions": {
						"type": "string",
						"title": "Instructions"
					},
					"_etag": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Etag"
					}
				},
				"type": "object",
				"required": ["instructions"],
				"title": "CurationInstructionsUpdate"
			},
			"DatasetCurationInstructions": {
				"properties": {
					"id": {
						"type": "string",
						"title": "Id"
					},
					"datasetName": {
						"type": "string",
						"title": "Datasetname"
					},
					"bucket": {
						"type": "string",
						"format": "uuid",
						"title": "Bucket"
					},
					"docType": {
						"type": "string",
						"title": "Doctype",
						"default": "curation-instructions"
					},
					"schemaVersion": {
						"type": "string",
						"title": "Schemaversion",
						"default": "v1"
					},
					"instructions": {
						"type": "string",
						"title": "Instructions"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"title": "Updatedat"
					},
					"updatedBy": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Updatedby"
					},
					"_etag": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Etag"
					}
				},
				"type": "object",
				"required": ["id", "datasetName", "instructions"],
				"title": "DatasetCurationInstructions",
				"description": "Dataset-level curation instructions document (schemaVersion v1).\n\nStored in the same Cosmos container as ground-truth items using MultiHash PK\n[/datasetName, /bucket] with bucket fixed to 0 and a stable id pattern\n\"curation-instructions|{datasetName}\"."
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		}
	}
}